+++ packages/ui/src/components/index.ts
@@ -1,30 +1,62 @@
 /**
- * Export only required set of components and utils from MUI, so we know what is used from the library and what is not
+ * Export only required set of components from MUI, so we know what is used from the library and what is not
  */
 
 export {
   Box,
-  type BoxProps,
   Button,
-  type ButtonProps,
   Stack,
-  type StackProps,
   Paper,
-  type PaperProps,
   Container,
-  type ContainerProps,
+  Card,
+  CardHeader,
+  CardContent,
   ToggleButton,
-  type ToggleButtonProps,
   ToggleButtonGroup,
-  type ToggleButtonGroupProps,
   Typography,
-  type TypographyProps,
   Link,
-  type LinkProps,
   Divider,
-  type DividerProps,
   Collapse,
-  type CollapseProps,
+  Tabs,
+  Tab,
+  Chip,
+  Avatar,
+  Drawer,
+  IconButton,
+  MenuList,
+  MenuItem,
+  ListItemIcon,
+  ListItemText,
+} from '@mui/material';
+
+export type {
+  BoxProps,
+  ButtonProps,
+  StackProps,
+  PaperProps,
+  ContainerProps,
+  CardProps,
+  CardHeaderProps,
+  CardContentProps,
+  ToggleButtonProps,
+  ToggleButtonGroupProps,
+  TypographyProps,
+  LinkProps,
+  DividerProps,
+  CollapseProps,
+  TabsProps,
+  TabProps,
+  ChipProps,
+  AvatarProps,
+  DrawerProps,
+  IconButtonProps,
+  MenuListProps,
+  MenuItemProps,
+  ListItemIconProps,
+  ListItemTextProps,
 } from '@mui/material';
 
 export * from './InfoTable';
+export * from './Dropdown';
+export * from './Loading';
+export * from './Logo';