+++ packages/ui/src/icons/index.ts
@@ -1,2 +1,18 @@
+/**
+ * Export only required set of icon from MUI, so we know what is used from the library and what is not.
+ * Add `*Icon` to the icon component name
+ */
+
+export {
+  Apps as AppsIcon,
+  MonetizationOn as MonetizationOnIcon,
+  Settings as SettingsIcon,
+  Category as CategoryIcon,
+  Menu as MenuIcon,
+  Close as CloseIcon,
+} from '@mui/icons-material';
+
+export type { IconProps } from '@mui/material';
+
 export * from './CereIcon';
 export * from './ChainIcon';