+++ packages/ui/src/theme/index.ts
@@ -1,4 +1,4 @@
-import { createTheme as createMuiTheme, alpha } from '@mui/material/styles';
+import { createTheme as createMuiTheme, alpha, Theme as MuiTheme } from '@mui/material/styles';
 
 declare module '@mui/material/styles/createPalette' {
   interface TypeText {
@@ -6,11 +6,13 @@ declare module '@mui/material/styles/createPalette' {
   }
 }
 
-export const createTheme = () => {
+export type Theme = MuiTheme;
+
+export const createTheme = (): Theme => {
   const theme = createMuiTheme({
     palette: {
       primary: {
-        main: '#8520E7',
+        main: '#733BF5',
       },
 
       secondary: {
@@ -18,6 +20,7 @@ export const createTheme = () => {
       },
 
       text: {
+        primary: '#131B32',
         secondary: '#717684',
         caption: '#A1A4AD',
       },
@@ -37,6 +40,14 @@ export const createTheme = () => {
         },
       },
 
+      MuiButtonBase: {
+        styleOverrides: {
+          root: {
+            borderRadius: '30px',
+          },
+        },
+      },
+
       MuiButton: {
         defaultProps: {
           disableElevation: true,
@@ -46,6 +57,10 @@ export const createTheme = () => {
           contained: {
             borderRadius: '30px',
           },
+
+          containedInherit: ({ theme }) => ({
+            backgroundColor: theme.palette.grey[100],
+          }),
         },
       },
 
@@ -56,6 +71,7 @@ export const createTheme = () => {
           },
         },
       },
+
       MuiToggleButtonGroup: {
         styleOverrides: {
           root: {
@@ -85,6 +101,44 @@ export const createTheme = () => {
               },
             },
           },
+        },
+      },
+
+      MuiTab: {
+        styleOverrides: {
+          root: ({ theme }) => ({
+            minHeight: '64px',
+
+            '& .MuiTab-iconWrapper:not(.Mui-selected *)': {
+              color: theme.palette.text.secondary,
+            },
+          }),
+
+          textColorPrimary: ({ theme }) => ({
+            color: theme.palette.text.primary,
+          }),
+        },
+      },
+
+      MuiMenuItem: {
+        styleOverrides: {
+          root: ({ theme }) => ({
+            '&.Mui-selected': {
+              color: theme.palette.primary.main,
+
+              '& .MuiListItemIcon-root': {
+                color: theme.palette.primary.main,
+              },
+            },
+          }),
+        },
+      },
+
+      MuiListItemIcon: {
+        styleOverrides: {
+          root: ({ theme }) => ({
+            color: theme.palette.text.secondary,
+          }),
         },
       },
     },