+++ src/components/index.ts
@@ -2,3 +2,4 @@ export * from './WalletWidget';
 export * from './PopupLayout';
 export * from './TransactionData';
 export * from './PriceRow';
+export * from './WalletLayout';