+++ src/stores/NetworkStore/NetworkStore.ts
@@ -1,14 +1,25 @@
 import { ChainConfig } from '@cere-wallet/communication';
 import { makeAutoObservable } from 'mobx';
 
-export class NetworkStore {
-  network?: ChainConfig;
+import { Wallet } from '../types';
+import { createSharedState } from '../sharedState';
 
-  constructor() {
+type SharedState = {
+  network?: ChainConfig;
+};
+
+export class NetworkStore {
+  private shared = createSharedState<SharedState>(`network.${this.wallet.instanceId}`, {});
+
+  constructor(private wallet: Wallet) {
     makeAutoObservable(this);
   }
 
-  setNetwork(config: ChainConfig) {
-    this.network = config;
+  get network() {
+    return this.shared.state.network;
+  }
+
+  set network(config: ChainConfig | undefined) {
+    this.shared.state.network = config;
   }
 }